Added first docs with simple example
