+++ docs/download_resources.jl
@@ -1,0 +1,15 @@
+# Download some assets necessary for docs/testing not stored in the repo
+import Downloads
+
+const directory = joinpath(@__DIR__, "src", "tutorials")
+mkpath(directory)
+
+for (file, url) in [
+        "periodic-rve.msh" => "https://raw.githubusercontent.com/Ferrite-FEM/Ferrite.jl/gh-pages/assets/periodic-rve.msh",
+        "periodic-rve-coarse.msh" => "https://raw.githubusercontent.com/Ferrite-FEM/Ferrite.jl/gh-pages/assets/periodic-rve-coarse.msh",
+    ]
+    afile = joinpath(directory, file)
+    if !isfile(afile)
+        Downloads.download(url, afile)
+    end
+end
