+++ docs/src/devdocs/cells.md
@@ -1,0 +1,22 @@
+# Cells
+
+## Working with interface cells
+
+```@docs
+FerriteInterfaceElements.get_interface_cell_shape
+FerriteInterfaceElements.get_sides_and_base_indices
+```
+
+## Inserting interface cells into a grid
+
+```@docs
+FerriteInterfaceElements.find_interfaces
+FerriteInterfaceElements.insert_duplicate_nodes!
+FerriteInterfaceElements.face_node_indices
+FerriteInterfaceElements.create_face_cell
+FerriteInterfaceElements.create_interface_cell
+FerriteInterfaceElements.create_interface_cells
+FerriteInterfaceElements.collect_cells_to_be_adapted
+FerriteInterfaceElements.adapt_cell
+FerriteInterfaceElements.adapt_cells!
+```
