+++ docs/src/devdocs/cellvalues.md
@@ -1,0 +1,5 @@
+# Cell Values
+
+```@docs
+FerriteInterfaceElements.get_base_value
+```
