+++ docs/src/devdocs/index.md
@@ -1,0 +1,13 @@
+# Developer documentation
+
+Here you can find some documentation of the internals of FerriteInterfaceElements.jl
+which are useful when developing the library.
+
+```@contents
+Depth = 1
+Pages = [    
+    "cells.md",
+    "interpolations.md",
+    "cellvalues.md",
+    ]
+```
