+++ docs/src/devdocs/interpolations.md
@@ -1,0 +1,7 @@
+# Interpolation
+
+```@docs
+FerriteInterfaceElements.get_n_dofs_on_side
+FerriteInterfaceElements.get_interface_index
+FerriteInterfaceElements.get_interface_dof_indices
+```
