+++ docs/src/reference/grid.md
@@ -1,0 +1,6 @@
+# Interface Cells
+
+```@docs
+InterfaceCell
+create_interface_cells!
+```
