+++ docs/src/reference/interpolations.md
@@ -1,0 +1,5 @@
+# Interpolation
+
+```@docs
+InterfaceCellInterpolation
+```
