+++ docs/src/tutorials/index.md
@@ -1,0 +1,16 @@
+# Tutorials
+
+On this page you find a collection of example problems as tutorials on how
+FerriteInterfaceElements can be used. The tutorials are organized in the same way as for
+[Ferrite](https://ferrite-fem.github.io/Ferrite.jl/stable/tutorials/).If you are new to
+Ferrite, you should start there to learn the basics.
+
+##### [Tutorial 1: Heat equation](heat_equation.md)
+
+This tutorial shows the basics steps using the heat equation as am example case for a
+scalar problem. A square domain with two phases is considered: a matrix with particle
+inclusions. First, the required cells at the interface are inserted, then all steps to
+colve the problem are presented. Finally, a visualization is generated using Makie.
+
+**Keywords**: scalar-valued solution, Dirichlet boundary conditions, 
+inserting interface cells, visualization using Makie.
