+++ test/runtests.jl
@@ -4,8 +4,6 @@ using Test
 
 @testset "FerriteInterfaceElements.jl" begin
     include("test_cells.jl")
-
-    include("test_grid.jl")
     
     include("test_interpolations.jl")
